refactor(cases): type invocation activities
