+++ tests/unit/test_case_comments_service.py
@@ -72,7 +72,7 @@ pytestmark = pytest.mark.usefixtures("db")
 def stub_case_duration_sync() -> Iterator[None]:
     with (
         patch(
-            "tracecat.cases.service.sync_case_duration",
+            "tracecat.cases.events.sync_case_duration",
             new=AsyncMock(return_value=True),
         ),
         patch("tracecat.cases.service.invoke_comment_agent_turns_after_commit"),
