+++ tracecat/cases/service.py
@@ -1,8 +1,7 @@
 import re
 import uuid
 from collections.abc import Sequence
-from datetime import UTC, datetime, timedelta
-from functools import partial
+from datetime import UTC, datetime
 from typing import Any, Literal
 from typing import cast as typing_cast
 
@@ -37,19 +36,15 @@ from tracecat.cases.dropdowns.schemas import (
     CaseDropdownValueRead,
 )
 from tracecat.cases.dropdowns.service import CaseDropdownValuesService
-from tracecat.cases.durations.materialization import sync_case_duration
 from tracecat.cases.durations.schemas import CaseDurationRead
-from tracecat.cases.durations.sync_queue import (
-    enqueue_case_duration_sync_after_commit,
-)
 from tracecat.cases.enums import (
-    CaseEventType,
     CasePriority,
     CaseSeverity,
     CaseStatus,
     CaseTaskStatus,
     MentionTargetType,
 )
+from tracecat.cases.events import CaseEventsService
 from tracecat.cases.mentions import MentionToken, parse_mentions
 from tracecat.cases.schemas import (
     AssigneeChangedEvent,
@@ -72,7 +67,6 @@ from tracecat.cases.schemas import (
     CaseTaskCreate,
     CaseTaskUpdate,
     CaseUpdate,
-    CaseViewedEvent,
     ClosedEvent,
     CommentCreatedEvent,
     CommentDeletedEvent,
@@ -111,7 +105,6 @@ from tracecat.db.models import (
     CaseDropdownDefinition,
     CaseDropdownOption,
     CaseDropdownValue,
-    CaseEvent,
     CaseFields,
     CaseTagLink,
     CaseTask,
@@ -220,8 +213,6 @@ def _enum_sort_rank(value: Any, ordered_values: Sequence[str]) -> int:
         return len(ordered_values)
 
 
-# Treat multiple views inside this window as a single "view" to avoid spam.
-CASE_VIEW_EVENT_DEDUP_WINDOW = timedelta(minutes=5)
 CASE_BATCH_LOCK_TIMEOUT = "5s"
 CASE_BATCH_LOCK_NOT_AVAILABLE_SQLSTATE = "55P03"
 PENDING_CASE_NUMBER = 0
@@ -2881,136 +2872,6 @@ class CaseCommentsService(BaseWorkspaceService):
             action="delete",
             comment_id=comment.id,
             data=delete_audit_data,
-        )
-
-
-class CaseEventsService(BaseWorkspaceService):
-    """Service for managing case events."""
-
-    service_name = "case_events"
-
-    def __init__(self, session: AsyncSession, role: Role | None = None):
-        super().__init__(session, role)
-
-    async def list_events(self, case: Case) -> Sequence[CaseEvent]:
-        """List all events for a case."""
-        statement = (
-            select(CaseEvent)
-            .where(CaseEvent.case_id == case.id)
-            # Order by creation time (newest first) and fall back to surrogate_id
-            # to ensure deterministic ordering when timestamps are equal.
-            .order_by(
-                CaseEvent.created_at.desc(),
-                CaseEvent.surrogate_id.desc(),
-            )
-        )
-        result = await self.session.execute(statement)
-        return result.scalars().all()
-
-    async def create_event(
-        self,
-        case: Case,
-        event: CaseEventVariant,
-        *,
-        publish_case_trigger: bool = True,
-        sync_durations: bool = True,
-        system_generated: bool = False,
-    ) -> CaseEvent:
-        """Create a new activity record for a case with variant-specific data.
-
-        Note: This method is non-committing. The caller is responsible for
-        wrapping operations in a transaction and committing once at the end
-        to preserve atomicity across multi-step updates.
-
-        Duration sync is queued after commit, with an inline fallback when
-        Redis publication fails.
-        """
-
-        db_event = CaseEvent(
-            workspace_id=self.workspace_id,
-            case_id=case.id,
-            type=event.type,
-            data=event.model_dump(exclude={"type"}, mode="json"),
-            user_id=None if system_generated else self.role.user_id,
-        )
-        self.session.add(db_event)
-        # Flush so that generated fields (e.g., id) are available if needed
-        await self.session.flush()
-
-        event_id = str(db_event.id)
-        event_type = (
-            db_event.type.value if hasattr(db_event.type, "value") else db_event.type
-        )
-        created_at = db_event.created_at or datetime.now(UTC)
-        case_id = str(case.id)
-        workspace_id = str(case.workspace_id)
-
-        if publish_case_trigger:
-
-            async def _publish_case_event() -> None:
-                await publish_case_event_payload(
-                    event_id=event_id,
-                    case_id=case_id,
-                    workspace_id=workspace_id,
-                    event_type=event_type,
-                    created_at=created_at,
-                )
-
-            AfterCommitQueue.of(self.session).add(_publish_case_event)
-
-        if sync_durations:
-            enqueue_case_duration_sync_after_commit(
-                self.session,
-                workspace_id=case.workspace_id,
-                case_id=case.id,
-                event_type=event_type,
-                reason="case_event",
-                inline_fallback=partial(
-                    sync_case_duration,
-                    case.workspace_id,
-                    case.id,
-                    event_types={event_type},
-                ),
-            )
-
-        return db_event
-
-    async def create_case_viewed_event(
-        self,
-        case: Case,
-        *,
-        dedupe_window: timedelta = CASE_VIEW_EVENT_DEDUP_WINDOW,
-    ) -> CaseEvent | None:
-        """Record a case viewed event if the current user hasn't viewed recently."""
-        if not self.role.user_id:
-            return None
-
-        now_utc = datetime.now(UTC)
-        stmt = (
-            select(CaseEvent)
-            .where(
-                CaseEvent.workspace_id == self.workspace_id,
-                CaseEvent.case_id == case.id,
-                CaseEvent.type == CaseEventType.CASE_VIEWED,
-                CaseEvent.user_id == self.role.user_id,
-            )
-            .order_by(CaseEvent.created_at.desc())
-            .limit(1)
-        )
-        result = await self.session.execute(stmt)
-        last_event = result.scalars().first()
-        if last_event:
-            last_created_at = last_event.created_at
-            if last_created_at.tzinfo is None:
-                if datetime.now() - last_created_at < dedupe_window:
-                    return None
-            else:
-                if now_utc - last_created_at < dedupe_window:
-                    return None
-
-        return await self.create_event(
-            case=case,
-            event=CaseViewedEvent(),
         )
